[fix] navigaitonbar UI 반영
